Fix: style and sweet alert
