+++ package.json
@@ -16,7 +16,9 @@
     "react-router-dom": "^6.7.0",
     "react-scripts": "^5.0.1",
     "sort-by": "^1.2.0",
-    "styled-components": "^5.3.6"
+    "styled-components": "^5.3.6",
+    "sweetalert": "^2.1.2",
+    "sweetalert2": "^11.7.1"
   },
   "scripts": {
     "start": "react-scripts start",
