+++ src/App.js
@@ -1,8 +1,8 @@
 import AuthProvider from "./contexts/AuthContext";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import SignIn from "./components/SignIn";
+import SignIn from "./components/Auth/SignIn.js";
 import { GlobalStyle } from "./globalStyles";
-import SignUp from "./components/SignUp";
+import SignUp from "./components/Auth/SignUp.js";
 import Home from "./components/Home/Home.js";
 import Cart from "./components/Cart";
 import Checkout from './components/Checkout/Checkout.js'
