+++ src/components/Auth/SignIn.js
@@ -1,27 +1,52 @@
-import styled from "styled-components";
-import axios from "axios";
 import { useNavigate } from "react-router-dom";
 import { useContext, useState } from "react";
+import { AuthContext } from "../../contexts/AuthContext";
+import Swal from "sweetalert2";
+import {Container, Form, Input, Button, GoTo } from "./styles.js";
+import { signInPost } from "../../services/auth.js";
 
 export default function SignIn() {
     const navigate = useNavigate();
+    const { setToken, userInfo, setUserInfo } = useContext(AuthContext);
     const [formInfo, setFormInfo] = useState({ email: '', password: '' })
 
     function handleForm(e) {
         setFormInfo({ ...formInfo, [e.target.name]: e.target.value });
-        console.log(formInfo);
     }
 
     function handleSubmit(e) {
         e.preventDefault();
-        console.log(formInfo);
-        const promise = axios.post(`http://localhost:5000/signin`, ({ ...formInfo }));
-        promise.then((res) => {
-            navigate("/")
-        });
-        promise.catch((error) => {
+        signInPost(formInfo).then((res) => {
+            setToken(res.data.token);
+            setUserInfo(res.data.user);
+
+            Swal.fire({
+                title: `Olá, ${userInfo.name}!`,
+                icon: 'success',
+            });
+            navigate("/");
+        })
+        .catch((error) => {
             console.log(error);
-            alert("Tente novamente");
+
+            if (error.response.status === 401) {
+                Swal.fire({
+                    title: 'Erro!',
+                    text: 'Você não está autorizado a fazer login, verifique seus dados ou realize um cadastro.',
+                    icon: 'error',
+                    confirmButtonText: 'Ok'
+                });
+            };
+
+            if (error.response.status === 500) {
+                Swal.fire({
+                    title: 'Erro',
+                    text: 'Aconteceu algo errado no nosso servidor, tente novamente mais tarde!',
+                    icon: 'error',
+                    cancelButtonText: 'Ok',
+                });
+            };
+
         })
     }
     return (<>
@@ -55,59 +80,3 @@ export default function SignIn() {
         </Container>
     </>)
 }
-const Container = styled.div`
-    background-color: #e5ebea;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    height: 668px;
-    margin: 0 auto;
-    font-family: "Open sans", sans-serif;
-    a {
-        text-decoration: none;
-    }
-`
-const Form = styled.form`
-background-color: #4c2a85;
-border-radius: 8px;
-display: flex;
-flex-direction: column;
-h1{
-    font-size: 30px;    
-    text-align: center;
-    margin-top: 20px;
-    margin-bottom: 20px;
-}
-`
-const Input = styled.input`
-border: none;
-border-radius: 4px;
-margin: 20px 20px;
-width: 380px;
-height: 45px;
-padding-left: 10px;
-font-size: 18px;
-color: #c6c6c6;
-`
-const Button = styled.button`
-background-color: #5f5aa2;
-    border: none;
-    border-radius: 4px;
-    width: 380px;
-    height: 40px;
-    padding-left: 10px;
-    font-size: 21px;
-    font-style: bold;
-    color: #fff;
-    line-height:35px;
-    margin: 20px 20px;
-`
-const GoTo = styled.p`
-color: #ffffff;
-text-align: center;
-font-size: 20px;
-line-height:30px;
-margin-bottom: 20px;
-`
