+++ src/contexts/AuthContext.js
@@ -6,9 +6,10 @@ export default function AuthProvider ({ children }) {
     const [token, setToken] = useState("");
     const [price, setPrice] = useState(0);
     const [items, setItems] = useState([]);
+    const [userInfo, setUserInfo] = useState([]);
 
     return (
-        <AuthContext.Provider value={{token, setToken, price, setPrice, items, setItems}}>
+        <AuthContext.Provider value={{token, setToken, price, setPrice, items, setItems, userInfo, setUserInfo}}>
             {children}
         </AuthContext.Provider>
     )
